requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
django-braces==1.11.0     # via django-oauth-toolkit
django-cors-middleware==1.3.1
django-oauth-toolkit==0.12.0
django==1.10.6
djangorestframework==3.6.4
first==2.0.1              # via pip-tools
idna==2.5                 # via requests
markdown==2.6.9
oauthlib==2.0.1           # via django-oauth-toolkit
pip-tools==1.8.0
requests==2.18.1
six==1.10.0               # via django-oauth-toolkit, pip-tools
urllib3==1.21.1           # via requests
waitress==1.0.2
whitenoise==3.3.0
